+++ firmware/src/RingBuffer.h
@@ -4,12 +4,13 @@
 #include "application.h"
 
 /**
- * Thread-safe(ish) circular buffer for audio samples.
+ * Thread-safe(ish) circular buffer for 16-bit audio samples.
  *
  * Improvements over the original SimpleRingBuffer:
  * - Tracks overflow count so callers know when samples were lost
  * - Uses volatile for fields accessed from ISR context
  * - Properly handles concurrent put() from ISR and get() from main loop
+ * - Stores 16-bit signed samples for better audio quality
  */
 class RingBuffer {
 public:
@@ -19,8 +20,8 @@ public:
   void init(unsigned int capacity);
   void destroy();
 
-  bool put(uint8_t value);
-  uint8_t get();
+  bool put(int16_t value);
+  int16_t get();
 
   void clear();
 
@@ -31,7 +32,7 @@ public:
   void resetOverflowCount();
 
 private:
-  uint8_t* _data;
+  int16_t* _data;
   unsigned int _capacity;
   volatile unsigned int _head;  // read position
   volatile unsigned int _tail;  // write position
